Move namers under cmdutils
These are only meant for interactive use in commands, not the API.
So cmdutils is the best place to keep these.
